backend/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "lxml>=6.0.2",
    "openai-agents>=0.6.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
